New snooze and radio on/off function:
When alarm is tripped:
If snooze is pressed, alarm is stopped and restarted after snooze time.

If radio toggle is pressed and alarm is running, radio is turned off and alarms cleared.

If radio is toggeled and alarm is paused for snooze, snooze is cleared and radio remains off.

When alarm is not tripped:

If radio is toggeled:
If radio is on due to snooze:
radio is turned off and snooze timer is cleared.
If no snooze is active:
Just toggle the radio on/off.